Made XY and XYZ properties be all doubles instead of ints

+++ GameCloudScape/src/game/worldmap/XY.java
@@ -2,10 +2,10 @@ package game.worldmap;
 
 public class XY
 {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 
-	public XY(int x, int y)
+	public XY(double x, double y)
 	{
 		this.x = x;
 		this.y = y;
@@ -16,12 +16,12 @@ public class XY
 		this.y = xyzPosition.getY();
 	}
 
-	public int getX()
+	public double getX()
 	{
 		return this.x;
 	}
 	
-	public int getY()
+	public double getY()
 	{
 		return this.y;
 	}
@@ -32,7 +32,7 @@ public class XY
         return 0;
     }
 	
-	public void add(int x, int y)
+	public void add(double x, double y)
 	{
 		this.x += x;
 		this.y += y;
